refactor(RecentProduct): DataSource 내부에서 Product와 매핑하는 것이 아닌 Repository 구현체 내부에서 수행하도록 수정
