+++ app/src/main/java/woowacourse/shopping/data/respository/recentproduct/source/local/RecentProductLocalDataSource.kt
@@ -1,9 +1,9 @@
 package woowacourse.shopping.data.respository.recentproduct.source.local
 
-import woowacouse.shopping.model.recentproduct.RecentProducts
+import woowacourse.shopping.data.model.RecentProductEntity
 
 interface RecentProductLocalDataSource {
     fun insertRecentProduct(productId: Long)
     fun deleteNotToday(today: String)
-    fun getAllRecentProducts(limit: Int, onSuccess: (RecentProducts) -> Unit)
+    fun getAllRecentProducts(limit: Int): List<RecentProductEntity>
 }
