+++ app/src/main/java/woowacourse/shopping/data/respository/recentproduct/source/local/RecentProductLocalDataSourceImpl.kt
@@ -9,9 +9,6 @@ import woowacourse.shopping.data.database.RecentProductHelper
 import woowacourse.shopping.data.database.getTableName
 import woowacourse.shopping.data.model.RecentProductEntity
 import woowacourse.shopping.data.model.Server
-import woowacourse.shopping.data.respository.product.source.remote.ProductRemoteDataSourceImpl
-import woowacouse.shopping.model.recentproduct.RecentProduct
-import woowacouse.shopping.model.recentproduct.RecentProducts
 import java.time.LocalDateTime
 
 class RecentProductLocalDataSourceImpl(
@@ -19,7 +16,6 @@ class RecentProductLocalDataSourceImpl(
     server: Server,
 ) : RecentProductLocalDataSource {
     private val db = RecentProductHelper(context).writableDatabase
-    private val productDataSource = ProductRemoteDataSourceImpl(server)
     private val tableName = getTableName(server)
 
     override fun insertRecentProduct(productId: Long) {
@@ -69,18 +65,8 @@ class RecentProductLocalDataSourceImpl(
         db.execSQL(sql)
     }
 
-    override fun getAllRecentProducts(limit: Int, onSuccess: (RecentProducts) -> Unit) {
-        productDataSource.requestDatas({}) { products ->
-            val recentProductEntities = getAllRecentProductsIds(limit)
-            val recentProducts = (
-                recentProductEntities.mapNotNull { recentProductEntity ->
-                    products.find { it.id == recentProductEntity.productId }?.let { product ->
-                        RecentProduct(recentProductEntity.id, product)
-                    }
-                }
-                ).toList()
-            onSuccess(RecentProducts(recentProducts))
-        }
+    override fun getAllRecentProducts(limit: Int): List<RecentProductEntity> {
+        return getAllRecentProductsIds(limit)
     }
 
     private fun getAllRecentProductsIds(limit: Int): List<RecentProductEntity> {
